fixed folder names

+++ scripts/process_docking/ligandWise/extract_energies.sh
@@ -5,52 +5,52 @@
 ############################################################
 Help() {
     echo "Syntax: extract_energies.sh [-h|d|f|o]"
-    echo "To save a log file and also print the status, run: extract_energies.sh -d \$DIRECTORY | tee -a \$LOGFILE"
+    echo "Requires an already processed DLG file (process_dlg.sh)."
+    echo "  The processed directory must be the same than "Processed DLG output directory" used by process_dlg.sh (-o flag)."
+    echo "  Also, requires the output of sort_pdb.sh."
     echo "Options:"
     echo "h     Print help"
-    echo "d     dlg files directory."
-    echo "o     Output directory."
+    echo "d     Ligand Name."
+    echo "i     Processed ligands' directory."
 }
 
-while getopts ":hd:o:" option; do
+while getopts ":hd:i:" option; do
     case $option in
         h)  # Print this help
             Help
             exit;;
         d)  # Enter the input directory
-            IPATH=$OPTARG;;
-        o)  # Output directory
-            OPATH=$OPTARG;;
+            LIGAND_NAME=$OPTARG;;
+        i)  # Output directory
+            PROCESSED_DIRECTORY=$OPTARG;;
         \?) # Invalid option
             echo "Error: Invalid option"
             exit;;
     esac
 done
 
-ALL_LIGAND_ENERGIES=${OPATH}/docking_scores.csv
-> $ALL_LIGAND_ENERGIES
+LIGAND_NAME=$(basename ${LIGAND_NAME} .dlg)
+LIGAND_PDB_PATH=${PROCESSED_DIRECTORY}/${LIGAND_NAME}/pdb/
 
-for DLG_FILE in "$IPATH"/*.dlg; do
+ENERGY_FILE="$LIGAND_PDB_PATH/docking_energies.txt"
 
-    LIGAND_PDBQT=$(basename $DLG_FILE .dlg)
-    LIGAND_NAME=$(basename $LIGAND_PDBQT .pdbqt)
-    LIGAND_PDB_PATH=${OPATH}/${LIGAND_NAME}/pdb/
-    ENERGY_FILE="$LIGAND_PDB_PATH/docking_energies.txt"
-    
-    sed -i 's/ /;/g' $ENERGY_FILE
+if [[ ! -f ${ENERGY_FILE} ]]
+then
+    echo "${ENERGY_FILE} not found."
+    echo "Did you run sort_pdb.sh?"
+    exit 1
+fi
 
-    TMP_FILE="$LIGAND_PDB_PATH/tmp.txt"
-    
-    seq $(cat $ENERGY_FILE | wc -l) | sed -E "s/.+/${LIGAND_NAME}/" > $TMP_FILE
+sed -i 's/ /;/g' $ENERGY_FILE
 
-    # Energies of single ligand
+TMP_FILE="$LIGAND_PDB_PATH/tmp.txt"
 
-    paste -d ';' $ENERGY_FILE $TMP_FILE > "$LIGAND_PDB_PATH/${LIGAND_NAME}_scores.csv"
+seq $(cat $ENERGY_FILE | wc -l) | sed -E "s/.+/${LIGAND_NAME}/" > $TMP_FILE
 
-    rm $TMP_FILE
-   
-    # All ligand energies
-    cat "$LIGAND_PDB_PATH/${LIGAND_NAME}_scores.csv" >> $ALL_LIGAND_ENERGIES
+# Energies of single ligand
 
-    
-done
+paste -d ';' $ENERGY_FILE $TMP_FILE > "${PROCESSED_DIRECTORY}/${LIGAND_NAME}/docking_scores.csv"
+rm $TMP_FILE
+
+# All ligand energies
+#cat "$LIGAND_PDB_PATH/${LIGAND_NAME}_scores.csv" >> $ALL_LIGAND_ENERGIES
